refactor(upstream): extract health/capability/discover modules (lab-kvji.12.6 steps 9-11)
- health.rs: circuit-breaker record_*/should_reprobe*/last_error/filter_collisions/
  upstream_count/status + 2 last_error tests (impl UpstreamPool block, methods stay pub).
- capability.rs: discover_capability_counts (pub(super), called from discover).
- discover.rs: discover_all_inner + discover_all{,_for_subject}{,_with_in_process_peers}
  + routable_upstream_peers (pub(super)). 430 LOC (<500).
Dropped redundant pub(crate) re-export of upstream_discovery_concurrency (no external
caller; discover imports it directly from helpers; helpers fn stays pub(crate)).
Behavior unchanged. 1423 green (1 pre-existing flake re-ran clean).

Bead: lab-kvji.12.6
