+++ crates/lab/src/dispatch/upstream/pool.rs
@@ -26,14 +26,17 @@ use crate::registry::{RegisteredService, ToolRegistry};
 use super::transport::websocket::{jitter_delay, reprobe_backoff};
 use super::types;
 use super::types::{
-    ToolExposurePolicy, UpstreamCapability, UpstreamEntry, UpstreamHealth, UpstreamRuntimeMetadata,
+    UpstreamCapability, UpstreamEntry, UpstreamHealth, UpstreamRuntimeMetadata,
     UpstreamRuntimeOwner, UpstreamTool, UpstreamToolExposureRow,
 };
 
+mod capability;
 mod connect;
 mod connect_stdio;
 mod connection;
+mod discover;
 mod entries;
+mod health;
 mod helpers;
 mod lifecycle;
 mod logging;
@@ -41,196 +44,19 @@ mod logging;
 mod testsupport;
 mod validate;
 
+pub(crate) use helpers::redact_resource_uri_for_logging;
 pub use helpers::{UpstreamCachedSummary, in_process_upstream_name};
-pub(crate) use helpers::{redact_resource_uri_for_logging, upstream_discovery_concurrency};
 // Leaf helpers used unqualified throughout the residual pool module and its
 // descendants. Glob-importing the child's `pub(super)` items keeps existing
 // call sites unchanged while the bodies live in the child modules.
 use connect::*;
 use connect_stdio::*;
+use discover::*;
 use entries::*;
 use helpers::*;
 use logging::*;
 use validate::*;
 
-/// Collect upstream peers for a capability in deterministic name order.
-async fn routable_upstream_peers(
-    pool: &UpstreamPool,
-    capability: UpstreamCapability,
-) -> Vec<(String, rmcp::service::Peer<RoleClient>)> {
-    let mut names: Vec<String> = {
-        let catalog = pool.catalog.read().await;
-        let mut names = match capability {
-            UpstreamCapability::Resources => {
-                let resource_names = pool.resource_upstreams.read().await;
-                resource_names
-                    .iter()
-                    .filter(|name| {
-                        catalog
-                            .get(*name)
-                            .is_some_and(|entry| entry.health_for(capability).is_routable())
-                    })
-                    .cloned()
-                    .collect::<Vec<_>>()
-            }
-            UpstreamCapability::Tools | UpstreamCapability::Prompts => catalog
-                .iter()
-                .filter(|(_, entry)| entry.health_for(capability).is_routable())
-                .map(|(name, _)| name.clone())
-                .collect::<Vec<_>>(),
-        };
-        names.sort_unstable();
-        names.dedup();
-        names
-    };
-
-    let acquire_started = Instant::now();
-    tracing::debug!(
-        surface = "dispatch",
-        service = "upstream.pool",
-        action = "upstream.acquire",
-        event = "start",
-        operation = "connection.acquire",
-        requested_operation = "capability.route",
-        capability = capability_name(capability),
-        requested_count = names.len(),
-        "upstream pool acquire start"
-    );
-    let connections = pool.connections.read().await;
-    let connection_count = connections.len();
-    let peers = names
-        .drain(..)
-        .filter_map(|name| connections.get(&name).map(|conn| (name, conn.peer.clone())))
-        .collect::<Vec<_>>();
-    drop(connections);
-    let pool_size = pool.catalog.read().await.len();
-    let elapsed_ms = acquire_started.elapsed().as_millis();
-    if peers.is_empty() {
-        tracing::warn!(
-            surface = "dispatch",
-            service = "upstream.pool",
-            action = "upstream.acquire",
-            event = "empty",
-            operation = "connection.acquire",
-            requested_operation = "capability.route",
-            capability = capability_name(capability),
-            elapsed_ms,
-            kind = "upstream_pool_empty",
-            pool_size,
-            connection_count,
-            "upstream pool acquire empty"
-        );
-    } else {
-        tracing::debug!(
-            surface = "dispatch",
-            service = "upstream.pool",
-            action = "upstream.acquire",
-            event = "finish",
-            operation = "connection.acquire",
-            requested_operation = "capability.route",
-            capability = capability_name(capability),
-            elapsed_ms,
-            pool_size,
-            connection_count,
-            acquired_count = peers.len(),
-            "upstream pool acquire finish"
-        );
-    }
-    peers
-}
-
-/// Returns true if the error represents a capability the upstream simply doesn't support
-/// (method not found / not implemented). These are healthy — the upstream just doesn't
-/// expose that capability, which is fine.
-async fn discover_capability_counts(
-    name: &str,
-    peer: &rmcp::service::Peer<RoleClient>,
-    proxy_resources: bool,
-    proxy_prompts: bool,
-) -> (
-    usize,
-    Option<String>,
-    UpstreamHealth,
-    usize,
-    Option<String>,
-    UpstreamHealth,
-) {
-    let (resource_count, resource_error, resource_health) = if proxy_resources {
-        tracing::info!(upstream = %name, capability = "resources", "starting upstream capability discovery");
-        match tokio::time::timeout(DISCOVERY_TIMEOUT, peer.list_resources(None)).await {
-            Ok(Ok(result)) => (result.resources.len(), None, UpstreamHealth::Healthy),
-            Ok(Err(ref error)) if is_capability_unsupported(error) => {
-                (0, None, UpstreamHealth::Healthy)
-            }
-            Ok(Err(error)) => (
-                0,
-                Some(format!("failed to list resources from upstream: {error}")),
-                UpstreamHealth::Unhealthy {
-                    consecutive_failures: 1,
-                },
-            ),
-            Err(_) => (
-                0,
-                Some(format!(
-                    "listing resources from upstream timed out after {}s",
-                    DISCOVERY_TIMEOUT.as_secs()
-                )),
-                UpstreamHealth::Unhealthy {
-                    consecutive_failures: 1,
-                },
-            ),
-        }
-    } else {
-        (0, None, UpstreamHealth::Healthy)
-    };
-
-    let (prompt_count, prompt_error, prompt_health) = if proxy_prompts {
-        tracing::info!(upstream = %name, capability = "prompts", "starting upstream capability discovery");
-        match tokio::time::timeout(DISCOVERY_TIMEOUT, peer.list_prompts(None)).await {
-            Ok(Ok(result)) => (result.prompts.len(), None, UpstreamHealth::Healthy),
-            Ok(Err(ref error)) if is_capability_unsupported(error) => {
-                (0, None, UpstreamHealth::Healthy)
-            }
-            Ok(Err(error)) => (
-                0,
-                Some(format!("failed to list prompts from upstream: {error}")),
-                UpstreamHealth::Unhealthy {
-                    consecutive_failures: 1,
-                },
-            ),
-            Err(_) => (
-                0,
-                Some(format!(
-                    "listing prompts from upstream timed out after {}s",
-                    DISCOVERY_TIMEOUT.as_secs()
-                )),
-                UpstreamHealth::Unhealthy {
-                    consecutive_failures: 1,
-                },
-            ),
-        }
-    } else {
-        (0, None, UpstreamHealth::Healthy)
-    };
-
-    if let Some(error) = &resource_error {
-        tracing::warn!(upstream = %name, error = %error, "failed to discover upstream resources");
-    }
-    if let Some(error) = &prompt_error {
-        tracing::warn!(upstream = %name, error = %error, "failed to discover upstream prompts");
-    }
-
-    (
-        resource_count,
-        resource_error,
-        resource_health,
-        prompt_count,
-        prompt_error,
-        prompt_health,
-    )
-}
-
-/// Merge upstream prompts deterministically and return the winning owner for each prompt.
 /// Upstream connection pool — holds live connections and discovered tool catalogs.
 #[derive(Clone)]
 pub struct UpstreamPool {
@@ -335,297 +161,6 @@ impl UpstreamPool {
     pub(super) fn with_request_timeout(mut self, timeout: Duration) -> Self {
         self.request_timeout = timeout;
         self
-    }
-
-    /// Connect to all configured upstreams and discover their tools.
-    ///
-    /// Each upstream gets a 15-second timeout. Failures are logged and the
-    /// upstream is marked unhealthy, but do not prevent other upstreams from
-    /// connecting.
-    #[allow(clippy::too_many_lines)]
-    async fn discover_all_inner(&self, configs: &[UpstreamConfig], oauth_subject: Option<&str>) {
-        if configs.is_empty() {
-            return;
-        }
-
-        // Validate name uniqueness and URI-safety before starting discovery.
-        let mut seen_names = std::collections::HashSet::new();
-        for config in configs {
-            if !seen_names.insert(&config.name) {
-                tracing::warn!(
-                    upstream = %config.name,
-                    "duplicate upstream name — skipping all but the first"
-                );
-            }
-            if !upstream_name_is_uri_safe(&config.name) {
-                tracing::warn!(
-                    upstream = %config.name,
-                    "upstream name contains URI-unsafe characters (/, ?, #) — skipping"
-                );
-            }
-        }
-
-        // Track which upstreams have resource/prompt proxying enabled.
-        let resource_names: Vec<String> = configs
-            .iter()
-            .filter(|c| c.enabled)
-            .filter(|c| c.proxy_resources)
-            .map(|c| c.name.clone())
-            .collect();
-        *self.resource_upstreams.write().await = resource_names;
-
-        let mut discovery_jobs = Vec::new();
-        let mut probe_configs = Vec::new();
-        let mut processed_names = std::collections::HashSet::new();
-        let oauth_client_cache = self.oauth_client_cache.clone();
-        let runtime_origin = self.runtime_origin.clone();
-        let runtime_owner = self.runtime_owner.clone();
-
-        for config in configs {
-            if !config.enabled {
-                continue;
-            }
-            // Skip duplicates (only process the first occurrence of each name).
-            if !processed_names.insert(&config.name) {
-                continue;
-            }
-            // Skip names with URI-unsafe characters.
-            if !upstream_name_is_uri_safe(&config.name) {
-                continue;
-            }
-            // OAuth HTTP upstreams require an explicit subject to select the
-            // correct token set. Subject-less discovery must skip them so one
-            // user's tool view is never cached globally by accident.
-            if config.oauth.is_some() && oauth_subject.is_none() {
-                tracing::info!(
-                    upstream = %config.name,
-                    transport = upstream_transport(config),
-                    target = %upstream_target_redacted(config),
-                    "skipping shared upstream discovery for subject-scoped OAuth upstream"
-                );
-                continue;
-            }
-            // Validate config
-            if let Err(msg) = validate_upstream_config(config) {
-                tracing::warn!(
-                    upstream = %config.name,
-                    "skipping upstream: {msg}"
-                );
-                continue;
-            }
-
-            let config = config.clone();
-            let subject = config
-                .oauth
-                .as_ref()
-                .and(oauth_subject)
-                .map(ToOwned::to_owned);
-            probe_configs.push(config.clone());
-            discovery_jobs.push((config, subject));
-        }
-
-        let discovery_concurrency = upstream_discovery_concurrency();
-        let mut futures = futures::stream::iter(discovery_jobs)
-            .map(|(config, subject)| {
-                let oauth_client_cache = oauth_client_cache.clone();
-                let runtime_origin = runtime_origin.clone();
-                let runtime_owner = runtime_owner.clone();
-                async move {
-                    let name = config.name.clone();
-                    match tokio::time::timeout(
-                        DISCOVERY_TIMEOUT,
-                        connect_upstream(
-                            &config,
-                            subject.as_deref(),
-                            oauth_client_cache.as_ref(),
-                            runtime_origin.as_deref(),
-                            runtime_owner.as_ref(),
-                        ),
-                    )
-                    .await
-                    {
-                        Ok(Ok((conn, tools))) => {
-                            let (
-                                resource_count,
-                                resource_last_error,
-                                resource_health,
-                                prompt_count,
-                                prompt_last_error,
-                                prompt_health,
-                            ) = discover_capability_counts(
-                                &name,
-                                &conn.peer,
-                                config.proxy_resources,
-                                config.proxy_prompts,
-                            )
-                            .await;
-                            tracing::info!(
-                                upstream = %name,
-                                transport = upstream_transport(&config),
-                                target = %upstream_target_redacted(&config),
-                                tool_count = tools.len(),
-                                resource_count,
-                                prompt_count,
-                                "upstream discovery succeeded"
-                            );
-                            Ok((
-                                name,
-                                config.expose_tools.clone(),
-                                conn,
-                                tools,
-                                resource_count,
-                                resource_last_error,
-                                resource_health,
-                                prompt_count,
-                                prompt_last_error,
-                                prompt_health,
-                            ))
-                        }
-                        Ok(Err(e)) => {
-                            let error = e.to_string();
-                            let kind = classify_upstream_error(&error);
-                            tracing::warn!(
-                                upstream = %name,
-                                transport = upstream_transport(&config),
-                                target = %upstream_target_redacted(&config),
-                                kind,
-                                error = %error,
-                                "upstream discovery failed"
-                            );
-                            Err((name, error))
-                        }
-                        Err(_) => {
-                            let error = format!(
-                                "upstream discovery timed out after {}s waiting for {} MCP list_tools response from {}",
-                                DISCOVERY_TIMEOUT.as_secs(),
-                                upstream_transport(&config),
-                                upstream_target_redacted(&config)
-                            );
-                            tracing::warn!(
-                                upstream = %name,
-                                transport = upstream_transport(&config),
-                                target = %upstream_target_redacted(&config),
-                                kind = "timeout",
-                                timeout_secs = DISCOVERY_TIMEOUT.as_secs(),
-                                "upstream discovery timed out"
-                            );
-                            Err((name, error))
-                        }
-                    }
-                }
-            })
-            .buffer_unordered(discovery_concurrency);
-
-        // Track all tool names across upstreams to detect duplicates.
-        let mut global_tool_names: HashMap<String, String> = HashMap::new();
-
-        while let Some(result) = futures.next().await {
-            match result {
-                Ok((
-                    name,
-                    expose_tools,
-                    conn,
-                    tools,
-                    resource_count,
-                    resource_last_error,
-                    resource_health,
-                    prompt_count,
-                    prompt_last_error,
-                    prompt_health,
-                )) => {
-                    let upstream_name: Arc<str> = Arc::from(name.as_str());
-                    let mut tool_map: HashMap<String, UpstreamTool> = HashMap::new();
-                    for tool in tools {
-                        let tool_name = tool.name.to_string();
-                        // Reject duplicate tool names across upstreams.
-                        if let Some(existing_upstream) = global_tool_names.get(&tool_name) {
-                            tracing::warn!(
-                                tool = %tool_name,
-                                upstream = %name,
-                                existing_upstream = %existing_upstream,
-                                "duplicate tool name across upstreams — skipping"
-                            );
-                            continue;
-                        }
-                        global_tool_names.insert(tool_name.clone(), name.clone());
-                        let (_, upstream_tool) = cached_upstream_tool(tool, &upstream_name);
-                        tool_map.insert(tool_name, upstream_tool);
-                    }
-
-                    let exposure_policy = resolve_exposure_policy(&name, expose_tools);
-
-                    let entry = UpstreamEntry {
-                        name: Arc::clone(&upstream_name),
-                        tools: tool_map,
-                        exposure_policy,
-                        prompt_count,
-                        resource_count,
-                        prompt_names: Vec::new(),
-                        resource_uris: Vec::new(),
-                        tool_health: UpstreamHealth::Healthy,
-                        prompt_health,
-                        resource_health,
-                        tool_unhealthy_since: None,
-                        prompt_unhealthy_since: (!prompt_health.is_routable()).then(Instant::now),
-                        resource_unhealthy_since: (!resource_health.is_routable())
-                            .then(Instant::now),
-                        tool_last_error: None,
-                        prompt_last_error,
-                        resource_last_error,
-                    };
-
-                    self.catalog.write().await.insert(name.clone(), entry);
-                    self.connections.write().await.insert(name, conn);
-                }
-                Err((name, error_message)) => {
-                    let entry = UpstreamEntry {
-                        name: Arc::from(name.as_str()),
-                        tools: HashMap::new(),
-                        exposure_policy: ToolExposurePolicy::All,
-                        prompt_count: 0,
-                        resource_count: 0,
-                        prompt_names: Vec::new(),
-                        resource_uris: Vec::new(),
-                        tool_health: UpstreamHealth::Unhealthy {
-                            consecutive_failures: 1,
-                        },
-                        prompt_health: UpstreamHealth::Unhealthy {
-                            consecutive_failures: 1,
-                        },
-                        resource_health: UpstreamHealth::Unhealthy {
-                            consecutive_failures: 1,
-                        },
-                        tool_unhealthy_since: Some(Instant::now()),
-                        prompt_unhealthy_since: Some(Instant::now()),
-                        resource_unhealthy_since: Some(Instant::now()),
-                        tool_last_error: Some(error_message.clone()),
-                        prompt_last_error: Some(error_message.clone()),
-                        resource_last_error: Some(error_message),
-                    };
-                    self.catalog.write().await.insert(name, entry);
-                }
-            }
-        }
-
-        for config in probe_configs {
-            self.ensure_probe_task(config);
-        }
-    }
-
-    /// Connect to non-OAuth upstreams and discover their tools.
-    ///
-    /// OAuth upstreams are intentionally skipped because they need a request or
-    /// gateway subject to select the right upstream token set.
-    pub async fn discover_all(&self, configs: &[UpstreamConfig]) {
-        self.discover_all_inner(configs, None).await;
-    }
-
-    /// Connect to all configured upstreams, using `subject` for OAuth upstreams.
-    ///
-    /// This is for gateway-owned discovery where the subject is an explicit
-    /// shared identity, not for subject-less startup discovery.
-    pub async fn discover_all_for_subject(&self, configs: &[UpstreamConfig], subject: &str) {
-        self.discover_all_inner(configs, Some(subject)).await;
     }
 
     fn ensure_probe_task(&self, config: UpstreamConfig) {
@@ -1162,25 +697,6 @@ impl UpstreamPool {
             entry.tools = tools;
             entry.exposure_policy = exposure_policy;
         }
-    }
-
-    pub async fn discover_all_with_in_process_peers(
-        &self,
-        configs: &[UpstreamConfig],
-        registry: &ToolRegistry,
-    ) {
-        self.discover_all(configs).await;
-        self.register_in_process_service_peers(registry).await;
-    }
-
-    pub async fn discover_all_for_subject_with_in_process_peers(
-        &self,
-        configs: &[UpstreamConfig],
-        subject: &str,
-        registry: &ToolRegistry,
-    ) {
-        self.discover_all_for_subject(configs, subject).await;
-        self.register_in_process_service_peers(registry).await;
     }
 
     pub async fn register_in_process_service_peers(&self, registry: &ToolRegistry) {
@@ -1800,172 +1316,6 @@ impl UpstreamPool {
     ///
     /// After [`CIRCUIT_BREAKER_THRESHOLD`] consecutive failures, the upstream
     /// is excluded from `list_tools` until a successful re-probe.
-    pub async fn record_failure(&self, upstream_name: &str, error: impl Into<String>) {
-        self.record_failure_for(upstream_name, UpstreamCapability::Tools, error)
-            .await;
-    }
-
-    /// Record a failure for a specific upstream capability, potentially marking it unhealthy.
-    ///
-    /// After [`CIRCUIT_BREAKER_THRESHOLD`] consecutive failures, the upstream
-    /// is excluded from the matching capability listing until a successful re-probe.
-    pub async fn record_failure_for(
-        &self,
-        upstream_name: &str,
-        capability: UpstreamCapability,
-        error: impl Into<String>,
-    ) {
-        let mut catalog = self.catalog.write().await;
-        if let Some(entry) = catalog.get_mut(upstream_name) {
-            let error = error.into();
-            let new_count = match entry.health_for(capability) {
-                UpstreamHealth::Healthy => 1,
-                UpstreamHealth::Unhealthy {
-                    consecutive_failures,
-                } => consecutive_failures + 1,
-            };
-            entry.set_health_for(
-                capability,
-                UpstreamHealth::Unhealthy {
-                    consecutive_failures: new_count,
-                },
-            );
-            if entry.unhealthy_since_for(capability).is_none() {
-                entry.set_unhealthy_since_for(capability, Some(Instant::now()));
-            }
-            entry.set_last_error_for(capability, Some(error.clone()));
-            if new_count >= types::CIRCUIT_BREAKER_THRESHOLD {
-                tracing::warn!(
-                    upstream = %upstream_name,
-                    capability = ?capability,
-                    consecutive_failures = new_count,
-                    error = %error,
-                    "circuit breaker open — upstream excluded from capability listing"
-                );
-            }
-        }
-    }
-
-    /// Record a success for an upstream capability, resetting the circuit breaker.
-    pub async fn record_success(&self, upstream_name: &str) {
-        self.record_success_for(upstream_name, UpstreamCapability::Tools)
-            .await;
-    }
-
-    /// Record a success for a specific upstream capability, resetting the circuit breaker.
-    pub async fn record_success_for(&self, upstream_name: &str, capability: UpstreamCapability) {
-        let mut catalog = self.catalog.write().await;
-        if let Some(entry) = catalog.get_mut(upstream_name) {
-            if !entry.health_for(capability).is_routable() {
-                tracing::info!(
-                    upstream = %upstream_name,
-                    capability = ?capability,
-                    "circuit breaker reset — upstream healthy"
-                );
-            }
-            entry.set_health_for(capability, UpstreamHealth::Healthy);
-            entry.set_unhealthy_since_for(capability, None);
-            entry.set_last_error_for(capability, None);
-        }
-    }
-
-    /// Return the most relevant last error for an upstream, if any capability has one.
-    pub async fn upstream_last_error(&self, upstream_name: &str) -> Option<String> {
-        let catalog = self.catalog.read().await;
-        let entry = catalog.get(upstream_name)?;
-        entry
-            .last_error_for(UpstreamCapability::Tools)
-            .or_else(|| entry.last_error_for(UpstreamCapability::Resources))
-            .or_else(|| entry.last_error_for(UpstreamCapability::Prompts))
-            .map(ToOwned::to_owned)
-    }
-
-    /// Return the last tools-capability error for an upstream, if any.
-    pub async fn upstream_tool_last_error(&self, upstream_name: &str) -> Option<String> {
-        let catalog = self.catalog.read().await;
-        let entry = catalog.get(upstream_name)?;
-        entry
-            .last_error_for(UpstreamCapability::Tools)
-            .map(ToOwned::to_owned)
-    }
-
-    #[cfg(test)]
-    pub async fn insert_entry_for_tests(&self, name: &str, entry: UpstreamEntry) {
-        self.catalog.write().await.insert(name.to_string(), entry);
-    }
-
-    /// Test-only: insert a fully-formed `UpstreamEntry` into the catalog.
-    pub async fn insert_entry_for_test(&self, name: &str, entry: UpstreamEntry) {
-        self.catalog.write().await.insert(name.to_string(), entry);
-    }
-
-    /// Check if an upstream capability is due for a re-probe.
-    #[allow(clippy::significant_drop_tightening)]
-    pub async fn should_reprobe(&self, upstream_name: &str) -> bool {
-        self.should_reprobe_for(upstream_name, UpstreamCapability::Tools)
-            .await
-    }
-
-    /// Check if a specific upstream capability is due for a re-probe.
-    #[allow(clippy::significant_drop_tightening)]
-    pub async fn should_reprobe_for(
-        &self,
-        upstream_name: &str,
-        capability: UpstreamCapability,
-    ) -> bool {
-        let catalog = self.catalog.read().await;
-        if let Some(entry) = catalog.get(upstream_name)
-            && entry.health_for(capability).is_open()
-            && let Some(since) = entry.unhealthy_since_for(capability)
-        {
-            return since.elapsed() >= types::REPROBE_INTERVAL;
-        }
-        false
-    }
-
-    /// Filter out upstream tools whose names collide with built-in service tools.
-    ///
-    /// Built-in lab services permanently take precedence. Upstream tools with
-    /// colliding names are dropped with a warning.
-    pub async fn filter_collisions(&self, builtin_names: &[&str]) {
-        let mut catalog = self.catalog.write().await;
-        for entry in catalog.values_mut() {
-            let collisions: Vec<String> = entry
-                .tools
-                .keys()
-                .filter(|name| builtin_names.contains(&name.as_str()))
-                .cloned()
-                .collect();
-            for name in &collisions {
-                tracing::warn!(
-                    upstream = %entry.name,
-                    tool = %name,
-                    "upstream tool name collides with built-in service — rejecting upstream tool"
-                );
-                entry.tools.remove(name);
-            }
-        }
-    }
-
-    /// Get the number of connected upstreams.
-    pub async fn upstream_count(&self) -> usize {
-        self.catalog.read().await.len()
-    }
-
-    #[cfg(test)]
-    pub async fn connection_count_for_tests(&self) -> usize {
-        self.connections.read().await.len()
-    }
-
-    /// Get names of all registered upstreams with their tool health status.
-    pub async fn upstream_status(&self) -> Vec<(String, UpstreamHealth)> {
-        let catalog = self.catalog.read().await;
-        catalog
-            .values()
-            .map(|e| (e.name.to_string(), e.tool_health))
-            .collect()
-    }
-
     /// Render the synthetic `lab://gateway/servers` document.
     ///
     /// Lists every registered upstream (regardless of health) with the
@@ -2788,6 +2138,7 @@ impl Default for UpstreamPool {
 
 #[cfg(test)]
 mod tests {
+    use super::super::types::ToolExposurePolicy;
     use super::testsupport::*;
     use super::*;
     use crate::mcp::logging::logging_level_rank;
@@ -3251,73 +2602,6 @@ mod tests {
 
         assert!(pool.find_tool("search_repos").await.is_some());
         assert!(pool.find_tool("delete_repo").await.is_none());
-    }
-
-    #[tokio::test]
-    async fn upstream_last_error_tracks_capability_failure_details() {
-        let pool = UpstreamPool::new();
-        let upstream_name: Arc<str> = Arc::from("github");
-        let entry = healthy_in_process_entry(Arc::clone(&upstream_name), HashMap::new());
-
-        pool.catalog
-            .write()
-            .await
-            .insert("github".to_string(), entry);
-
-        pool.record_failure_for(
-            "github",
-            UpstreamCapability::Resources,
-            "resource listing returned 401 unauthorized",
-        )
-        .await;
-
-        assert_eq!(
-            pool.upstream_last_error("github").await.as_deref(),
-            Some("resource listing returned 401 unauthorized")
-        );
-
-        pool.record_success_for("github", UpstreamCapability::Resources)
-            .await;
-        assert_eq!(pool.upstream_last_error("github").await, None);
-    }
-
-    #[tokio::test]
-    async fn upstream_tool_last_error_ignores_non_tool_failures() {
-        let pool = UpstreamPool::new();
-        let upstream_name: Arc<str> = Arc::from("github");
-        let entry = healthy_in_process_entry(Arc::clone(&upstream_name), HashMap::new());
-
-        pool.catalog
-            .write()
-            .await
-            .insert("github".to_string(), entry);
-
-        pool.record_failure_for(
-            "github",
-            UpstreamCapability::Resources,
-            "resource listing returned 401 unauthorized",
-        )
-        .await;
-        pool.record_failure_for(
-            "github",
-            UpstreamCapability::Prompts,
-            "prompt listing returned 501 unsupported",
-        )
-        .await;
-
-        assert_eq!(pool.upstream_tool_last_error("github").await, None);
-
-        pool.record_failure_for(
-            "github",
-            UpstreamCapability::Tools,
-            "tool listing returned 500 internal error",
-        )
-        .await;
-
-        assert_eq!(
-            pool.upstream_tool_last_error("github").await.as_deref(),
-            Some("tool listing returned 500 internal error")
-        );
     }
 
     #[tokio::test]
